More changes

+++ samples/samples/conftest.py
@@ -31,6 +31,10 @@ INSTANCE_CREATION_TIMEOUT = 560  # seconds
 
 OPERATION_TIMEOUT_SECONDS = 120  # seconds
 
+CREATE_INSTANCE = False
+
+INSTANCE_ID_DEFAULT = "google-cloud-python-systest"
+
 retry_429 = retry.RetryErrors(exceptions.ResourceExhausted, delay=15)
 
 
@@ -92,7 +96,9 @@ def cleanup_old_instances(spanner_client):
 @pytest.fixture(scope="module")
 def instance_id():
     """Unique id for the instance used in samples."""
-    return f"test-instance-{uuid.uuid4().hex[:10]}"
+    if CREATE_INSTANCE:
+        return f"test-instance-{uuid.uuid4().hex[:10]}"
+    return INSTANCE_ID_DEFAULT
 
 
 @pytest.fixture(scope="module")
@@ -121,31 +127,36 @@ def sample_instance(
     instance_config,
     sample_name,
 ):
-    sample_instance = spanner_client.instance(
-        instance_id,
-        instance_config,
-        labels={
-            "cloud_spanner_samples": "true",
-            "sample_name": sample_name,
-            "created": str(int(time.time())),
-        },
-    )
-    op = retry_429(sample_instance.create)()
-    op.result(INSTANCE_CREATION_TIMEOUT)  # block until completion
+    if CREATE_INSTANCE:
+        sample_instance = spanner_client.instance(
+            instance_id,
+            instance_config,
+            labels={
+                "cloud_spanner_samples": "true",
+                "sample_name": sample_name,
+                "created": str(int(time.time())),
+            },
+        )
+        op = retry_429(sample_instance.create)()
+        op.result(INSTANCE_CREATION_TIMEOUT)  # block until completion
 
-    # Eventual consistency check
-    retry_found = retry.RetryResult(bool)
-    retry_found(sample_instance.exists)()
+        # Eventual consistency check
+        retry_found = retry.RetryResult(bool)
+        retry_found(sample_instance.exists)()
+    else:
+        sample_instance = spanner_client.instance(instance_id)
+        sample_instance.reload()
 
     yield sample_instance
 
-    for database_pb in sample_instance.list_databases():
-        database.Database.from_pb(database_pb, sample_instance).drop()
+    if CREATE_INSTANCE:
+        for database_pb in sample_instance.list_databases():
+            database.Database.from_pb(database_pb, sample_instance).drop()
 
-    for backup_pb in sample_instance.list_backups():
-        backup.Backup.from_pb(backup_pb, sample_instance).delete()
+        for backup_pb in sample_instance.list_backups():
+            backup.Backup.from_pb(backup_pb, sample_instance).delete()
 
-    sample_instance.delete()
+        sample_instance.delete()
 
 
 @pytest.fixture(scope="module")
